jwt: Use "long long" for json integers if jansson supports it
On win64 (of course), a long is only 4-bytes, which could lead to
overflows for things like "exp" claims.

We key off of JSON_INTEGER_IS_LONG_LONG in jansson_config.h to sync
with that library. We still avoid exposing jansson.h in jwt.h, so
that's good.

This might break ABI on win64. I'm not sure. It almost certainly
breaks ABI on 32-bit systems where the anonymous union in jwt_value_t
would only have been 4-bytes.

On win64 the union was 8-bytes anyway because of the pointers. Now,
however, we wont be casting claims to 32-bits in this case.

So...bump the SO_VERSION, or not? Time will tell.

Fixes #220

+++ libjwt/jwt-setget.c
@@ -46,7 +46,7 @@ static jwt_value_error_t jwt_get_int(json_t *which, jwt_value_t *jval)
 	else if (!json_is_integer(val))
 		return jval->error = JWT_VALUE_ERR_TYPE;
 
-	jval->int_val = (long)json_integer_value(val);
+	jval->int_val = json_integer_value(val);
 
 	return jval->error;
 }
